requirements.txt
google-generativeai
openai
IPython
google-colab